notification/requirements.txt
aio-pika==9.5.4
aiormq==6.8.1
annotated-types==0.7.0
exceptiongroup==1.2.2
idna==3.10
multidict==6.1.0
pamqp==3.3.0
propcache==0.3.0
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
python-dotenv==1.0.1
typing_extensions==4.12.2
yarl==1.18.3
